Grading UI and versioned reviews by teachers
- assignment_controller now sends a list of review assignments by version
- ct.js now contains a frontend for showing reviews (defaults to reviewing the
  first version)

+++ web/app/models/path_ref.rb
@@ -32,7 +32,11 @@ class PathRef < ActiveRecord::Base
   def contents
     if Rails.env.development?
       p = File.expand_path(self.path, user_repo.path)
-      File.open(p).gets(nil)
+      if File.exists?(p)
+        File.open(p).gets(nil)
+      else
+        user_repo.lookup_file_head(self.path)
+      end
     else
       user_repo.lookup_file_head(self.path)
     end
